openzwave: fix gcc7 compile and remove bad patch
Patch 903-no-examples.patch removes MinOZW binary. This is wrong since
install section of openzwave is still using it. We can either remove the
patch or remove install section for openzwave. It seems people are using
this binary so we remove the patch.

After removing the patch the following issue appears with gcc7:
cc1plus: error: -Wformat-security ignored without -Wformat

This is resolved by adding the following to Makefile
TARGET_CPPFLAGS+=-Wno-format -Wno-format-security

Signed-off-by: Marko Ratkaj <marko.ratkaj@sartura.hr>