autossh: add an 'enabled' option within the uci configuration file
In a tool like this one, you really want an option to establish if the service
should start or not by default on boot time, especially when its configuration
file has to be customized by the user.

In the configuration file, the new 'enabled' option is setted to '0' by default
since the configuration provided by default will not be the one finally used.

In the init script, the new 'enabled' option is setted to '1' by default in
order to support the previous configuration file behaviour.

Signed-off-by: Adrià Llaudet <adria.llaudet@gmail.com>