bmx7-dnsupdate: add package
Makes it easy to address bmx7 shorids and hostnames with ping, ssh, etc.
Both are stored in /tmp/hosts/ and so loaded by dnsmasq.
Instead of using a cron job which triggers the bmx7 deamon every so and
so, inotifywait is used to see creation or deletion of originators in
/var/run/bmx7/json/originators/

Signed-off-by: Paul Spooren <spooren@informatik.uni-leipzig.de>